example/requirements/base.txt
django-germanium>=2.0.5
coverage==4.0.2
https://github.com/druids/django-pyston/tarball/BetterPermissions#egg=django-pyston-2.9.4
https://github.com/matllubos/django-is-core/tarball/BetterPermissionsModelv2#egg=django-is-core
factory_boy>=2.11.1
-e ../
